requirements.txt
aiocache==0.11.1
aiofiles==0.8.0
aiohttp==3.7.4.post0
APScheduler==3.8.1
async-timeout==3.0.1
attrs==21.2.0
autopep8==2.0.1
beautifulsoup4==4.11.1
bs4==0.0.1
chardet==4.0.0
colorthief==0.2.1
discord.py @ git+https://github.com/Rapptz/discord.py@06df0d7931385a5b8efc1dc6bea79cc389b22808
dnspython==2.2.1
e==1.4.5
emoji-country-flag==1.3.0
expiringdict==1.2.1
fold-to-ascii==1.0.2.post1
humanfriendly==10.0
humanize==3.12.0
idna==3.3
mongoengine==0.23.1
msgpack==1.0.2
multidict==5.2.0
numpy==1.24.2
opencv-python==4.7.0.68
packaging==23.0
Pillow==9.3.0
psutil==5.8.0
pycodestyle==2.10.0
pymongo==3.12.1
pytesseract==0.3.10
python-dotenv==0.19.1
pytimeparse==1.1.8
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
soupsieve==2.3.2.post1
toml==0.10.2
tomli==2.0.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
ujson==4.2.0
yarl==1.7.0
chatgpt @ git+https://github.com/itsnebulalol/chatgpt@main
